Idented a chkroot() a little better and added a reminder in # TODO.
As of this commit, my, your, our code will follow a code-style of: code wrote inside a function will need to be spaced with two spaces from the editor "column", and code inside it will need two more spaces (see if-else-fi and loops for having a better understanding of what im saying here).

Added a reminder to write the remove() functions and finish others @ #TODO (line 40).